Print all answers for a set of letters

print_answers is a cheater function which identifies all the possible answers for a set of given letters.

print_answers(
  game_letters = NULL,
  central = NULL,
  num_letters = 7,
  min_word_length = 4,
  dictionary = "normal",
  obscenities = FALSE,
  game = NULL,
  state = NULL
)

Arguments

game_letters

The letters to identify words from. These should be a character string (e.g., "polemic"). (Required, if game is not provided.)

central

The central (that is, required) letter (optional). If not provided, will be randomly chosen from among the game_letters.

num_letters

The number of letters for the game; defaults to 7, and should be between 6 and 10.

min_word_length

Expected length of words. Defaults to 4 letters, but can be between 2 and 6.

dictionary

Character string. Choice of how detailed of a dictionary to use. Can be any of "slim", "broad", or "normal". Defaults to "normal", which is recommended.

obscenities

Logical. Should obscenities be included? Defaults to FALSE.

game

A holoalphabetic game from play_game or create_game. If included, no other parameters are required.

state

State of a holoalphabetic game; provided from current gameplay.

Value

When called directly (or through gameplay), this prints the answers to all words possible.

Details

These words may or may not be the same as those identified with other word sets; you can modify by changing the chosen dictionary. This function is called by play_game when choosing to get all answers (with "[a]" entered instead of a guess).

This function will always check whether you really want to see the answers, requiring a "y" to continue.

See also

Examples

print_answers(game = create_game("craking"))
#> Ready to see the full word list? Type a y if so!: ? 
print_answers("balmntu")
#> Ready to see the full word list? Type a y if so!: ?