Prints hints to help get all the answers

print_hint_list is a helper function which identifies the number of words for the given set of letters, and (if desired) identifies what letters (two or three) words begin with.

print_hint_list(
  game_letters = NULL,
  central = NULL,
  num_letters = 7,
  min_word_length = 4,
  dictionary = "normal",
  obscenities = FALSE,
  game = NULL,
  hint = NULL
)

Arguments

game_letters

The letters to identify words from. These should be a character string (e.g., "polemic"). (Required, if game is not provided.)

central

The central (that is, required) letter (optional). If not provided, will be randomly chosen from among the game_letters.

num_letters

The number of letters for the game; defaults to 7, and should be between 6 and 10.

min_word_length

Expected length of words. Defaults to 4 letters, but can be between 2 and 6.

dictionary

Character string. Choice of how detailed of a dictionary to use. Can be any of "slim", "broad", or "normal". Defaults to "normal", which is recommended.

obscenities

Logical. Should obscenities be included? Defaults to FALSE.

game

A holoalphabetic game from play_game or create_game. If included, no other parameters are required.

state

State of a holoalphabetic game; provided from current gameplay.

Value

When called directly (or through gameplay), this prints hints to help you find all answers.

Details

These words may or may not be the same as those identified with other word sets; you can make minor changes by changing the chosen dictionary. This function is called by play_game when choosing to get hints (with "[h]" entered instead of a guess). You can choose to get hints of the first two of three letters, or none.

See also

Examples

print_hint_list(game = create_game("craking"))
#> Words: 65, Pangrams: 3 (1 perfect). 
#>      
#>        4  5  6  7  8  9 10 Sum
#>   a    3  2  4  1  1  1  1  13
#>   c    1  5  5  2  2  -  -  15
#>   g    4  1  -  4  1  -  -  10
#>   i    -  -  -  1  -  -  1   2
#>   k    1  1  -  -  -  -  1   3
#>   n    3  1  1  1  -  -  -   6
#>   r    5  1  4  5  1  -  -  16
#>   Sum 17 11 14 14  5  1  3  65
#> Would you like to see the first two (type '2') or three (type '3') letters of the words? 
#>  If neither, just hit enter.
#> ? 
print_hint_list("balmntu")
#> Words: 48, Pangram: 1 (0 perfect). 
#>      
#>        4  5  6  7  8 Sum
#>   a    4  2  3  -  2  11
#>   b    3  1  3  1  -   8
#>   l    3  1  -  -  -   4
#>   m    4  4  4  -  -  12
#>   n    1  1  -  -  -   2
#>   t    4  3  -  -  1   8
#>   u    1  1  1  -  -   3
#>   Sum 20 13 11  1  3  48
#> Would you like to see the first two (type '2') or three (type '3') letters of the words? 
#>  If neither, just hit enter.
#> ?